add in jwt

+++ src/main/java/com/example/demo/controller/UserController.java
@@ -4,23 +4,41 @@ import com.example.demo.config.CommonResult;
 import com.example.demo.dto.*;
 import com.example.demo.models.User;
 import com.example.demo.service.impl.UserServiceImpl;
+import com.example.demo.utils.JwtTokenProvider;
 import com.example.demo.vo.ListUserVo;
 import com.example.demo.vo.LoginVo;
 import com.example.demo.vo.UserVo;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import cn.hutool.crypto.digest.Digester;
+import io.jsonwebtoken.security.Keys;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import cn.hutool.crypto.digest.DigestAlgorithm;
 
+
+import java.security.Key;
+import java.util.Date;
 import java.util.List;
 
 @CrossOrigin(origins = "http://localhost:3000")
 @RequestMapping("/user")
 @RestController
 public class UserController {
+
     @Autowired
     private UserServiceImpl userServiceImpl;
+    @Autowired
+    private JwtTokenProvider provider;
     @PostMapping("/add")
     public CommonResult<UserVo> addUser(@RequestBody NewUserDto user) {
         User newUser = userServiceImpl.addUser(user);
+        if(newUser == null) {
+            return CommonResult.failed(404,"User exist already");
+        }
         UserVo userResponse = UserVo.builder()
                 .id(newUser.getId())
                 .name(newUser.getName())
@@ -58,6 +76,10 @@ public class UserController {
 
     @DeleteMapping(path="/delete") // Map ONLY DELETE Requests
     public CommonResult<UserVo> deleteUser (@RequestBody DeletedUserDto deletedUser) {
+        long userId = provider.validateToken(deletedUser.getToken());
+        if(userId != deletedUser.getId()) {
+            return CommonResult.failed(401,"User not authorised");
+        }
         if(userServiceImpl.existsById(deletedUser.getId())){ //if user exists
             User user = userServiceImpl.deleteUser(deletedUser.getId());
             UserVo userResponse = UserVo.builder()
@@ -102,6 +124,10 @@ public class UserController {
 
     @PutMapping("/update")
     public CommonResult<UserVo> updateUser(@RequestBody UserDto user) {
+        long userId = provider.validateToken(user.getToken());
+        if(userId != user.getId()) {
+            return CommonResult.failed(401,"User not authorised");
+        }
         User updatedUser = userServiceImpl.updateUser(user);
         UserVo userResponse = UserVo.builder()
                 .id(updatedUser.getId())
@@ -115,11 +141,16 @@ public class UserController {
 
     @PostMapping("/login")
     public CommonResult<LoginVo> getById(@RequestBody LoginUserDto loginUser) {
-        boolean userFound = userServiceImpl.findUser(loginUser.getName(), loginUser.getPassword());
+        User userFound = userServiceImpl.findUser(loginUser.getName(), loginUser.getPassword());
+        String token = null;
+        if(userFound != null) { //if login successful
+             token = provider.generateToken(userFound, Keys.secretKeyFor(SignatureAlgorithm.HS256));
+        }
         LoginVo loginResponse = LoginVo.builder()
-                .login(userFound)
+                .loginUser(userFound)
+                .token(token)
                 .build();
-        if(userFound) {
+        if(userFound!=null) {
             return CommonResult.success(loginResponse, "Login success");
         } else {
             return CommonResult.failed(404,"Login unsuccessful");
